requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.9.1
    # via django
beautifulsoup4==4.13.4
    # via libpysal
certifi==2025.7.14
    # via
    #   pyogrio
    #   pyproj
    #   requests
charset-normalizer==3.4.2
    # via requests
contourpy==1.3.2
    # via matplotlib
coverage==7.9.2
    # via -r requirements.in
cycler==0.12.1
    # via matplotlib
django==5.2.4
    # via
    #   -r requirements.in
    #   django-ninja
    #   django-stubs
    #   django-stubs-ext
django-ninja==1.4.3
    # via -r requirements.in
django-stubs==5.2.1
    # via -r requirements.in
django-stubs-ext==5.2.1
    # via django-stubs
fonttools==4.58.5
    # via matplotlib
geopandas==1.1.1
    # via libpysal
idna==3.10
    # via requests
inequality==1.1.2
    # via -r requirements.in
joblib==1.5.1
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
libpysal==4.13.0
    # via inequality
matplotlib==3.10.3
    # via inequality
mypy==1.16.1
    # via django-stubs
mypy-extensions==1.1.0
    # via mypy
numpy==2.3.1
    # via
    #   contourpy
    #   geopandas
    #   inequality
    #   libpysal
    #   matplotlib
    #   pandas
    #   pyogrio
    #   scikit-learn
    #   scipy
    #   shapely
packaging==25.0
    # via
    #   geopandas
    #   libpysal
    #   matplotlib
    #   pyogrio
pandas==2.3.1
    # via
    #   geopandas
    #   inequality
    #   libpysal
pathspec==0.12.1
    # via mypy
pillow==11.3.0
    # via matplotlib
platformdirs==4.3.8
    # via libpysal
pydantic==2.11.7
    # via django-ninja
pydantic-core==2.33.2
    # via pydantic
pyogrio==0.11.0
    # via geopandas
pyparsing==3.2.3
    # via matplotlib
pyproj==3.7.1
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
requests==2.32.4
    # via libpysal
ruff==0.12.3
    # via -r requirements.in
scikit-learn==1.7.0
    # via libpysal
scipy==1.16.0
    # via
    #   inequality
    #   libpysal
    #   scikit-learn
shapely==2.1.1
    # via
    #   geopandas
    #   libpysal
six==1.17.0
    # via python-dateutil
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
threadpoolctl==3.6.0
    # via scikit-learn
types-pyyaml==6.0.12.20250516
    # via django-stubs
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
